requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
streamlit==1.39.0
PyPDF2==3.0.1
pytesseract==0.3.13
pillow==10.4.0
chromadb==0.5.11
openai==1.51.2
sentence-transformers==3.2.0
httpx==0.27.2